requirements.txt
azure-ai-formrecognizer==3.3.2
streamlit==1.29.0
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
pdf2image==1.17.0